extract test helpers in separate file
