+++ tests/unit/parser/songLinesFactory.spec.js
@@ -5,6 +5,12 @@ import _ from 'lodash';
 
 import songLinesFactory from '../../../src/parser/songLinesFactory';
 
+import {
+	getLyricLine,
+	getChordLine,
+	getSectionLine,
+	getEmptyLine,
+} from './songLinesFactory-helpers.spec';
 import parseChordLine from '../../../src/parser/parseChordLine';
 import parseSectionLabel from '../../../src/parser/parseSectionLabel';
 import parseTimeSignature from '../../../src/parser/parseTimeSignature';
@@ -1140,202 +1146,5 @@ verseContent2
 		input.split('\n').forEach(songLines.addLine);
 
 		expect(songLines.asArray()).toEqual(expected);
-	});
-});
-
-/** ====================== **/
-/**   Helpers functions    **/
-/** ====================== **/
-
-function getSectionLine(
-	sectionStr,
-	id,
-	index,
-	{ fromCopy = false, fromRepeat = false } = {}
-) {
-	const sectionLine = {
-		type: 'sectionLabel',
-		string: sectionStr,
-		index: index,
-		indexWithoutRepeats: index,
-		model: parseSectionLabel(sectionStr),
-		id,
-	};
-	if (fromCopy) {
-		sectionLine.isFromSectionCopy = true;
-	}
-	if (fromRepeat) {
-		sectionLine.isFromSectionRepeat = true;
-	}
-	return sectionLine;
-}
-
-function getLine(
-	type,
-	lineString,
-	{ fromCopy = false, fromRepeat = false } = {}
-) {
-	const line = {
-		type,
-		string: lineString,
-	};
-	if (type !== 'emptyLine') {
-		line.model = lineString;
-	}
-	if (fromCopy) {
-		line.isFromSectionCopy = true;
-	}
-	if (fromRepeat) {
-		line.isFromSectionRepeat = true;
-	}
-	return line;
-}
-
-function getLyricLine(lineStr, options) {
-	return getLine('lyric', lineStr, options);
-}
-
-function getChordLine(lineStr, options) {
-	return getLine('chord', lineStr, options);
-}
-
-function getEmptyLine(options) {
-	return getLine('emptyLine', '', options);
-}
-
-describe('Helpers functions', () => {
-	describe.each([
-		['base', undefined, undefined, getLyricLine('lyricContent')],
-		['no copy, no repeat', false, false, getLyricLine('lyricContent')],
-		[
-			'copy, no repeat',
-			true,
-			false,
-			{
-				type: 'lyric',
-				model: 'lyricContent',
-				string: 'lyricContent',
-				isFromSectionCopy: true,
-			},
-		],
-		[
-			'no copy, repeat',
-			false,
-			true,
-			{
-				type: 'lyric',
-				model: 'lyricContent',
-				string: 'lyricContent',
-				isFromSectionRepeat: true,
-			},
-		],
-		[
-			'copy, repeat',
-			true,
-			true,
-			{
-				type: 'lyric',
-				model: 'lyricContent',
-				string: 'lyricContent',
-				isFromSectionCopy: true,
-				isFromSectionRepeat: true,
-			},
-		],
-	])('Lyric line: %s', (title, fromCopy, fromRepeat, expected) => {
-		test('getLyricLine', () => {
-			const actual = getLyricLine('lyricContent', {
-				fromCopy,
-				fromRepeat,
-			});
-			expect(actual).toEqual(expected);
-		});
-	});
-
-	describe.each([
-		[
-			'base',
-			undefined,
-			undefined,
-			{ type: 'chord', model: 'chordContent', string: 'chordContent' },
-		],
-		[
-			'no copy, no repeat',
-			false,
-			false,
-			{ type: 'chord', model: 'chordContent', string: 'chordContent' },
-		],
-		[
-			'copy, no repeat',
-			true,
-			false,
-			{
-				type: 'chord',
-				model: 'chordContent',
-				string: 'chordContent',
-				isFromSectionCopy: true,
-			},
-		],
-		[
-			'no copy, repeat',
-			false,
-			true,
-			{
-				type: 'chord',
-				model: 'chordContent',
-				string: 'chordContent',
-				isFromSectionRepeat: true,
-			},
-		],
-		[
-			'copy, repeat',
-			true,
-			true,
-			{
-				type: 'chord',
-				model: 'chordContent',
-				string: 'chordContent',
-				isFromSectionCopy: true,
-				isFromSectionRepeat: true,
-			},
-		],
-	])('Chord line: %s', (title, fromCopy, fromRepeat, expected) => {
-		test('getChordLine()', () => {
-			const actual = getChordLine('chordContent', {
-				fromCopy,
-				fromRepeat,
-			});
-			expect(actual).toEqual(expected);
-		});
-	});
-
-	describe.each([
-		['base', undefined, undefined, { type: 'emptyLine', string: '' }],
-		['no copy, no repeat', false, false, { type: 'emptyLine', string: '' }],
-		[
-			'copy, no repeat',
-			true,
-			false,
-			{ type: 'emptyLine', string: '', isFromSectionCopy: true },
-		],
-		['no copy, repeat', false, true, getEmptyLine({ fromRepeat: true })],
-		[
-			'copy, repeat',
-			true,
-			true,
-			{
-				type: 'emptyLine',
-				string: '',
-				isFromSectionCopy: true,
-				isFromSectionRepeat: true,
-			},
-		],
-	])('Empty line: %s', (title, fromCopy, fromRepeat, expected) => {
-		test('getEmptyLine()', () => {
-			const actual = getEmptyLine({
-				fromCopy,
-				fromRepeat,
-			});
-			expect(actual).toEqual(expected);
-		});
 	});
 });
